+++ gcs/CHANGES.md
@@ -1,8 +1,14 @@
 ### 2.2.0 - 2020-XX-XX
 
-1. Delete deprecated methods.
+1.  Delete deprecated methods.
 
 1.  Update all dependencies to latest versions.
+
+1.  Add support for Cloud Storage objects CSEK encryption:
+
+        fs.gs.encryption.algorithm (not set by default)
+        fs.gs.encryption.key (not set by default)
+        fs.gs.encryption.key.hash (not set by default)
 
 ### 2.1.1 - 2020-03-11